[fix] Updated with changes #576
Fixes #576
