+++ openwisp_monitoring/device/apps.py
@@ -5,7 +5,6 @@ from django.conf import settings
 from django.core.cache import cache
 from django.db.models import Count
 from django.db.models.signals import post_delete, post_save
-from django.dispatch import receiver
 from django.urls import reverse_lazy
 from django.utils.translation import gettext_lazy as _
 from swapper import get_model_name, load_model
@@ -53,26 +52,19 @@ class DeviceMonitoringConfig(AppConfig):
         self.add_connection_ignore_notification_reasons()
 
     def connect_check_signals(self):
+        from django.db.models.signals import post_delete, post_save
         from swapper import load_model
 
         Check = load_model('check', 'Check')
         DeviceMonitoring = load_model('device_monitoring', 'DeviceMonitoring')
 
-        @receiver(post_save, sender=Check)
-        def check_post_save_receiver(sender, instance, **kwargs):
-            DeviceMonitoring.handle_critical_metric(instance.metric)
-
-        @receiver(post_delete, sender=Check)
-        def check_post_delete_receiver(sender, instance, **kwargs):
-            DeviceMonitoring.handle_critical_metric(instance.metric)
-
         post_save.connect(
-            check_post_save_receiver,
+            DeviceMonitoring.handle_critical_metric,
             sender=Check,
             dispatch_uid='check_post_save_receiver',
         )
         post_delete.connect(
-            check_post_delete_receiver,
+            DeviceMonitoring.handle_critical_metric,
             sender=Check,
             dispatch_uid='check_post_delete_receiver',
         )
